+++ R/write_config.R
@@ -39,7 +39,6 @@
 #' @examples
 #' x <- setup_SuperConfig(
 #'   dat_training_path = "~/Data/iris.csv",
-#'   preprocessor_config = setup_Preprocessor(remove_duplicates = TRUE),
 #'   hyperparameters = setup_LightRF(),
 #'   outer_resampling_config = setup_Resampler(),
 #'   question = "Can we tell iris species apart given their measurements?",